DijkstraGeneric line 220 with short cast fixed

+++ src/main/java/org/graphast/query/route/shortestpath/dijkstra/DijkstraGeneric.java
@@ -217,7 +217,7 @@ public class DijkstraGeneric {
 	    int waitingTime, timeToService;
 	    
 	    shortestDistances.put(v, 0);
-	    QueueEntry e = new QueueEntry(v, (short) 0);
+	    QueueEntry e = new QueueEntry(v, 0);
         unsettledNodes.add(e);
         
         while ((e = unsettledNodes.poll()) != null){
